+++ Tools/TestRunnerShared/UIScriptContext/UIScriptController.h
@@ -202,7 +202,7 @@ public:
     virtual JSObjectRef calendarType() const { notImplemented(); return nullptr; }
     virtual void setDefaultCalendarType(JSStringRef calendarIdentifier) { notImplemented(); }
     virtual JSObjectRef inputViewBounds() const { notImplemented(); return nullptr; }
-    virtual void activateDataListSuggestion(long, JSValueRef) { notImplemented(); }
+    virtual void activateDataListSuggestion(unsigned, JSValueRef) { notImplemented(); }
 
     // Share Sheet
